cheri: drivers: usb: Fix compile
Align structures with pointers to a pointer and add uintptr_t casts.

+++ drivers/usb/atm/ueagle-atm.c
@@ -241,7 +241,7 @@ enum {
 #define UEA_IS_PREFIRM(x) \
 	(!((x)->driver_info & PSTFIRM))
 #define UEA_CHIP_VERSION(x) \
-	((x)->driver_info & 0xf)
+	((unsigned long)(x)->driver_info & 0xf)
 
 #define IS_ISDN(x) \
 	((x)->annex & ANNEXB)
